refactor: move Client/Server utility functions to separate headers
